requirements.txt
aioflask==0.4.0
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
blinker==1.7.0
cachetools==4.2.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
db-dtypes==1.2.0
eventkit==1.0.3
Flask==3.0.2
forex-python==1.8
frozenlist==1.4.1
google-api-core==2.17.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.17.2
google-cloud-bigquery-storage==2.24.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
greenletio==0.11.0
grpcio==1.60.1
grpcio-status==1.60.1
gunicorn==21.2.0
h11==0.14.0
ib-insync==0.9.86
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pandas==2.2.0
pandas-gbq==0.21.0
proto-plus==1.23.0
protobuf==4.25.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydata-google-auth==1.8.2
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.19.2
six==1.16.0
structlog==24.1.0
tzdata==2024.1
urllib3==2.2.0
uvicorn==0.27.1
Werkzeug==3.0.1
yarl==1.9.4
